Imported SFX .WAV Files to Audio Folder
I hvae imported the .wav files to the audio folder. In the sound design channel in discord there will be a screenshot of the file directory plan for the folder.

Only one .wav file failed to import into the game build. Should be ready to import after a day. Just needs to be converted again into a .wav file.