AttackComponent
Removed AttackType from the Attack function. This is because we can handle all of this by having the weapon set in the class itself.
Updated the Attack function to take a TSubClassOf<AWeapon>, this means only a weapon can be passed into the function
If the weaponType isn't set on the Weapon Class it will default to Null and if it is used at runtime will display a debug message to the screen.