Debugging Functionality
Seems like we need to find a more reliable way to assign the tag option to the checkpoint system. As the current way removes the tag option of the checkpoint when its loaded up the game level again.

I have, for now, made an event end overlap on the bonfires, and moved the player starts to be within the bounds of the collision box, so we get the checkpoint tag regardless.